exercicio aula49
