--- semana17/aula49-filtros-ordenacao-paginacao/src/types/user.ts
@@ -1,9 +1,0 @@
-import { recipe } from "./recipe"
-
-export type user = {
-   id: string,
-   name: string,
-   email: string,
-   recipes: recipe[],
-   password: string
-}
