+++ semana17/template-aula49-filtros-ordenacao-paginacao (1)/template-aula49-filtros-ordenacao-paginacao/__MACOSX/._tsconfig.json
